+++ src/javascript/api/components/tabs.js
@@ -116,7 +116,6 @@ module.exports = {
                     screen: current_screen.id
                   },
                   function () {
-                    editingTab = id
                     const tab_created_event = new CustomEvent('tab_created', {
                       detail: {
                         tab: tab